+++ sites-list.txt
@@ -30,4 +30,3 @@ https://www.nfe.fazenda.gov.br/portal/principal.aspx
 
 Vivo
 http://vivo.com.br/
-
