+++ src/Components/CheckWindowSize/styles.css
@@ -19,6 +19,7 @@
   0% {
     transform: rotate(0deg);
   }
+
   75%,
   100% {
     transform: rotate(90deg);
