+++ src/Components/LoadingSpinner/index.js
@@ -1,7 +1,7 @@
 import "./styles.css";
 
 export function loadingSpinnerHTML() {
-  return `
+    return `
         <div class="vwo-spinner-container">
             <div class="vwo-spinner"></div>
         </div>
