+++ src/InitialConfigs/Images.js
@@ -3,9 +3,9 @@ import { loadModal } from "../Components/Modal";
 let working = false;
 
 export function ImagesConfig() {
-  if(!working) createLinksContainer();
+  if (!working) createLinksContainer();
   working = true;
-  
+
   document.querySelectorAll("img").forEach((img) => {
     if (img.clientWidth + img.clientHeight > 100) {
       img.classList.add("vwo-img-ocr");
